Remove dota2vods.net link and fix GG link
Removed the /clan from the end of the GG link within the extension as this takes users to a login gateway. Removing it takes them to the list without having to login or anything. dota2vods.net is no longer actively around so a bit pointless to have an empty broken tab.